+++ src/components/WithdrawPage.tsx
@@ -9,8 +9,8 @@ import { AppDispatch } from "../state"
 import { AppState } from "../state"
 import { BigNumber } from "@ethersproject/bignumber"
 import ConfirmTransaction from "./ConfirmTransaction"
-import DeadlineField from "./DeadlineField"
-import GasField from "./GasField"
+import DeadlineField from "./material/DeadlineField"
+import GasField from "./material/GasField"
 import { HistoricalPoolDataType } from "../hooks/useHistoricalPoolData"
 import InfiniteApprovalField from "./material/InfiniteApprovalField"
 import Modal from "./Modal"
@@ -21,7 +21,7 @@ import { PayloadAction } from "@reduxjs/toolkit"
 import PoolInfoCard from "./PoolInfoCard"
 import RadioButton from "./RadioButton"
 import ReviewWithdraw from "./ReviewWithdraw"
-import SlippageField from "./SlippageField"
+import SlippageField from "./material/SlippageField"
 import TokenInput from "./TokenInput"
 import { WithdrawFormState } from "../hooks/useWithdrawFormState"
 import classNames from "classnames"
